chore(interface): document behaviour of overlapping subscriptions
The specifications say that the broker MAY deduplicate messages
on overlapping subscriptions. In result, some do, and some don't.

It is now, by documentation, left to the user of this library to
handle overlapping subscriptions properly, and depending on their
broker, they may receive one or more times the same message when
the subscriptions overlap.