fix(connection): check every 100ms, not every 100us, for an update
Doing it every 100us consumes a lot of CPU, and was not the
intention.

Additionally, increase the read interval to 100ms, as 10ms is using
quite a lot of CPU for not really a good reason other than reacting
quickly on error situations. That really is not the right balance
to strike.